09.11.night .

+++ src/controllers/qnaController.js
@@ -60,11 +60,21 @@ export const postUploadQ = async (req, res) => {
 
 export const getEditQ = async (req, res) => {
     const { id } = req.params;
+    
     const question = await Question.findById(id); // 질문 데이터 object를 찾아 가져오는 model
-
     if (!question) {
         return res.status(404).render("404", { pageTitle: "Question not found." });
     }
+
+    // 프론트에서는 링크를 숨겼지만 백엔드에서 추가로 보호해줘야함
+    const {
+        user: { _id },
+      } = req.session;
+
+    if (String(question.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+    
     return res.render("qna/editQ", {pageTitle : "Edit Question", question});
 };
 
@@ -72,24 +82,50 @@ export const postEditQ = async (req, res) => {
     // Question : 우리가 만든 질문 model
     // question : 데이터베이스에서 검색한 질문 object
     const { id } = req.params;
-    const { title, content, hashtags} = req.body;
-    const question = await Question.exists({ _id: id }); // 질문 데이터 존재 여부만 판단하는 model
-
-    if (!question) {
-        return res.status(404).render("404", { pageTitle: "Question not found." });
-    }
+    const { title, content, hashtags } = req.body;
     await Question.findByIdAndUpdate(id, {
         title,
         content,
         hashtags: Question.formatHashtags(hashtags),
-      });
+    });
+
+    // 질문 데이터 존재 여부만 판단하는 model
+    const question = await Question.exists({ _id: id });
+    if (!question) {
+        return res.status(404).render("404", { pageTitle: "Question not found." });
+    }
+    
+    // 프론트에서는 링크를 숨겼지만 백엔드에서 추가로 보호해줘야함
+    const {
+        user: { _id },
+    } = req.session;
+
+    if (String(question.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+
     return res.redirect("/qna");
 };
 
 export const deleteQ = async (req, res) => {
     const { id } = req.params;
-    await Question.findByIdAndDelete(id);
 
+    // 질문 데이터 존재 여부만 판단하는 model
+    const question = await Question.findById(id);
+    if (!question) {
+        return res.status(404).render("404", { pageTitle: "Question not found." });
+    }
+    
+    // 프론트에서는 링크를 숨겼지만 백엔드에서 추가로 보호해줘야함
+    const {
+        user: { _id },
+    } = req.session;
+
+    if (String(question.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+
+    await Question.findByIdAndDelete(id);
     return res.redirect("/qna");
   };
